+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/lifecycle/CatchupProcessListener.java
@@ -1,8 +1,0 @@
-package uk.gov.justice.services.event.sourcing.subscription.lifecycle;
-
-public interface CatchupProcessListener {
-
-    default void onCatchupStarted(@SuppressWarnings("unused") final CatchupStartedEvent catchupStartedEvent) {}
-
-    default void onCatchupCompleted(@SuppressWarnings("unused") final CatchupCompletedEvent catchupCompletedEvent) {}
-}